- convert GoldScreen from StatefulWidget to StatelessWidget
This commit refactors the `GoldScreen` widget from a `StatefulWidget` to a `StatelessWidget`. The commented-out `initState` method, which was previously used to fetch company data, is removed as it's no longer necessary with the widget being stateless. Data fetching is likely handled elsewhere, potentially within the `GoldCubit` or triggered by another mechanism.